Attempt to fix music crashes FIXED bot sometimes crashing at the beginning of a song (?) BUMPED to v3.2.1

+++ index.js
@@ -153,6 +153,7 @@ client.player.events.on("playerStart", async function(queue, track) {
 });
 // WHEN SONG STOPS PLAYING (END, STOP OR SKIP)
 client.player.events.on("playerFinish", async function(queue, track) {
+    if (!client.player.musicControlsMessage.fetch()) return console.log("> DEBUG - MUSIC crash caught on index.js:156");
     await client.player.musicControlsMessage.edit({ components: [] });
     setTimeout(async function(message){
         await message.delete();
